refactor: Move balanceTVL data fetching to parent component
This will facilitate Overview tab with aggregate tvls
